# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "An API for users to register/login and view different products.\n                    Registered users can add products to their cart and checkout to create \n                    their order. Registered users with admin role can add, edit and delete the products."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get all products. Admins only",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched list of users",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Users found"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Update user by id. Admins only",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the user to update",
            "example": 1
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User fields to update",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "Updated username"
                },
                "firstname": {
                  "type": "string",
                  "example": "Updated firstname"
                },
                "lastname": {
                  "type": "string",
                  "example": "Updated lastname"
                },
                "email": {
                  "type": "string",
                  "example": "Updated email"
                },
                "address": {
                  "type": "string",
                  "example": "Updated address"
                },
                "phone": {
                  "type": "string",
                  "example": "Updated phone"
                },
                "role_id": {
                  "type": "string",
                  "example": "Updated role_id"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "User updated"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad request — missing required fields.",
            "schema": {
              "$ref": "#components/schemas/MissingParameters"
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#components/schemas/ResourceNotFound"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "firstname": {
          "type": "string",
          "example": "Example firstname"
        },
        "lastname": {
          "type": "string",
          "example": "Example lastname"
        },
        "username": {
          "type": "string",
          "example": "Example username"
        },
        "email": {
          "type": "string",
          "example": "example@gmail.com"
        },
        "address": {
          "type": "string",
          "example": "Example address"
        },
        "phone": {
          "type": "string",
          "example": "12345678"
        },
        "membership": {
          "type": "string",
          "example": "bronze"
        },
        "role": {
          "type": "string",
          "example": "User"
        }
      },
...
    }
  }
}